Refactoring and making code shorter

+++ bus-health/src/main/java/org/aoju/bus/health/builtin/hardware/NetworkIF.java
@@ -113,7 +113,7 @@ public interface NetworkIF {
      * This value is set when the {@link NetworkIF} is
      * instantiated and may not be up to date.
      */
-    int getMTU();
+    long getMTU();
 
     /**
      * The Media Access Control (MAC) address.
